# be/docs/swagger.yaml
basePath: /api
definitions:
  dto.ApproveLeaveRequest:
    properties:
      rejection_reason:
        type: string
      status:
        $ref: '#/definitions/model.LeaveStatus'
    required:
    - status
    type: object
...
  dto.CompanyResponse:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      logo:
        type: string
      name:
        type: string
      npwp:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.CreateLeaveRequest:
    properties:
      attachment:
        type: string
      employee_id:
        type: string
      end_date:
        type: string
      leave_type:
        $ref: '#/definitions/model.LeaveType'
      reason:
        type: string
      start_date:
        type: string
      total_days:
        type: integer
    required:
    - employee_id
    - end_date
    - leave_type
    - reason
    - start_date
    - total_days
    type: object
...
  dto.DepartmentResponse:
    properties:
      company:
        $ref: '#/definitions/dto.CompanyResponse'
      company_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
  dto.EmployeeResponse:
    properties:
      bank_account:
        type: string
      bank_name:
        type: string
      birth_date:
        type: string
      birth_place:
        type: string
      blood_type:
        type: string
      bpjs_kes_no:
        type: string
      bpjs_tk_no:
        type: string
      company:
        $ref: '#/definitions/dto.CompanyResponse'
      company_id:
        type: string
      created_at:
        type: string
      department:
        $ref: '#/definitions/dto.DepartmentResponse'
      department_id:
        type: string
      employee_number:
        type: string
      employee_status:
        $ref: '#/definitions/model.EmployeeStatus'
      gender:
        type: string
      id:
        type: string
      join_date:
        type: string
      last_education:
        type: string
      marital_status:
        type: string
      nik:
        type: string
      npwp:
        type: string
      position:
        $ref: '#/definitions/dto.PositionResponse'
      position_id:
        type: string
      religion:
        type: string
      resign_date:
        type: string
      shift:
        $ref: '#/definitions/dto.ShiftResponse'
      shift_id:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
      user_id:
        type: string
    type: object
...
  dto.LeaveResponse:
    properties:
      approved_at:
        type: string
      approved_by:
        type: string
      approver:
        $ref: '#/definitions/dto.UserResponse'
      attachment:
        type: string
      created_at:
        type: string
      employee:
        $ref: '#/definitions/dto.EmployeeResponse'
      employee_id:
        type: string
      end_date:
        type: string
      id:
        type: string
      leave_type:
        $ref: '#/definitions/model.LeaveType'
      reason:
        type: string
      rejection_reason:
        type: string
      start_date:
        type: string
      status:
        $ref: '#/definitions/model.LeaveStatus'
      total_days:
        type: integer
      updated_at:
        type: string
    type: object
...
  dto.MenuAccessResponse:
    properties:
      menu_keys:
        items:
          type: string
        type: array
      user_email:
        type: string
      user_id:
        type: string
      user_name:
        type: string
      user_role:
        type: string
    type: object
  dto.OrgDepartmentNode:
    properties:
      id:
        type: string
      name:
        type: string
      positions:
        items:
          $ref: '#/definitions/dto.OrgPositionNode'
        type: array
    type: object
  dto.OrgEmployeeNode:
    properties:
      employee_number:
        type: string
      id:
        type: string
      user_name:
        type: string
    type: object
  dto.OrgPositionNode:
    properties:
      employees:
        items:
          $ref: '#/definitions/dto.OrgEmployeeNode'
        type: array
      id:
        type: string
      name:
        type: string
    type: object
  dto.OrgStructureResponse:
    properties:
      company_id:
        type: string
      company_name:
        type: string
      departments:
        items:
          $ref: '#/definitions/dto.OrgDepartmentNode'
        type: array
    type: object
...
  dto.PositionResponse:
    properties:
      base_salary:
        type: number
      company:
        $ref: '#/definitions/dto.CompanyResponse'
      company_id:
        type: string
      created_at:
        type: string
      department:
        $ref: '#/definitions/dto.DepartmentResponse'
      department_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
  dto.SetMenuAccessRequest:
    properties:
      menu_keys:
        items:
          type: string
        type: array
      user_id:
        type: string
    required:
    - menu_keys
    - user_id
    type: object
  dto.ShiftResponse:
    properties:
      company:
        $ref: '#/definitions/dto.CompanyResponse'
      company_id:
        type: string
      created_at:
        type: string
      end_time:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      start_time:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.UpdateLeaveRequest:
    properties:
      attachment:
        type: string
      end_date:
        type: string
      leave_type:
        $ref: '#/definitions/model.LeaveType'
      reason:
        type: string
      start_date:
        type: string
      total_days:
        type: integer
    type: object
...
  dto.UserMenuKeysResponse:
    properties:
      menu_keys:
        items:
          type: string
        type: array
    type: object
  dto.UserResponse:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/model.Role'
      updated_at:
        type: string
    type: object
...
  model.EmployeeStatus:
    enum:
    - tetap
    - kontrak
    - probation
    type: string
    x-enum-varnames:
    - StatusTetap
    - StatusKontrak
    - StatusProbation
  model.LeaveStatus:
    enum:
    - pending
    - approved
    - rejected
    type: string
    x-enum-varnames:
    - LeaveStatusPending
    - LeaveStatusApproved
    - LeaveStatusRejected
  model.LeaveType:
    enum:
    - cuti_tahunan
    - cuti_sakit
    - cuti_melahirkan
    - cuti_besar
    - izin
    - dinas_luar
    type: string
    x-enum-varnames:
    - LeaveCutiTahunan
    - LeaveCutiSakit
    - LeaveCutiMelahirkan
    - LeaveCutiBesar
    - LeaveIzin
    - LeaveDinasLuar
...
  model.Role:
    enum:
    - admin
    - hr
    - employee
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleHR
    - RoleEmployee
  response.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@hris.com
    name: API Support
  description: Human Resource Information System API with authentication and comprehensive
    HRIS modules
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: HRIS API
  version: "1.0"
paths:
...
  /leaves:
    get:
      description: Retrieve all leave requests, optionally filtered by employee or
        status
      parameters:
      - description: Filter by employee ID
        in: query
        name: employee_id
        type: string
      - description: Filter by status (e.g. pending)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Leaves retrieved
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.LeaveResponse'
                  type: array
              type: object
        "500":
          description: Failed to fetch leaves
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get all leave requests
      tags:
      - Leaves
    post:
      consumes:
      - application/json
      description: Submit a new leave request
      parameters:
      - description: Leave request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateLeaveRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Leave request created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LeaveResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create a leave request
      tags:
      - Leaves
  /leaves/{id}:
    delete:
      description: Delete a leave request by ID
      parameters:
      - description: Leave ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Leave request deleted
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to delete
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete a leave request
      tags:
      - Leaves
    get:
      description: Retrieve a leave request by its ID
      parameters:
      - description: Leave ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Leave retrieved
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LeaveResponse'
              type: object
        "404":
          description: Leave not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get leave request by ID
      tags:
      - Leaves
    put:
      consumes:
      - application/json
      description: Update an existing leave request by ID
      parameters:
      - description: Leave ID
        in: path
        name: id
        required: true
        type: string
      - description: Leave request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateLeaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Leave request updated
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LeaveResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update a leave request
      tags:
      - Leaves
  /leaves/{id}/approve:
    put:
      consumes:
      - application/json
      description: Approve or reject a pending leave request (Admin/HR only)
      parameters:
      - description: Leave ID
        in: path
        name: id
        required: true
        type: string
      - description: Approval data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ApproveLeaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Leave request updated
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LeaveResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Approve or reject a leave request
      tags:
      - Leaves
  /menu-access:
    get:
      description: Retrieve all user menu access mappings (Admin only)
      produces:
      - application/json
      responses:
        "200":
          description: All menu access retrieved
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.MenuAccessResponse'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get all menu access mappings
      tags:
      - Menu Access
    post:
      consumes:
      - application/json
      description: Set the menu keys accessible to a specific user (Admin only)
      parameters:
      - description: Menu access data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SetMenuAccessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Menu access updated
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Set user menu access
      tags:
      - Menu Access
  /menu-access/{user_id}:
    delete:
      description: Delete custom menu access for a user, reverting to role defaults
        (Admin only)
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Menu access reset to defaults
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: user_id is required
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Reset user menu access
      tags:
      - Menu Access
  /menu-access/me:
    get:
      description: Retrieve the list of menu keys accessible to the currently authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: Menu access retrieved
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserMenuKeysResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get my menu access
      tags:
      - Menu Access
  /organization/structure:
    get:
      description: Retrieve the hierarchical organization structure for a company
        (departments → positions → employees)
      parameters:
      - description: Company ID
        in: query
        name: company_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Organization structure retrieved
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.OrgStructureResponse'
              type: object
        "400":
          description: company_id is required
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Company not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get organization structure
      tags:
      - Organization
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"